+++ stdnet/orm/globals.py
@@ -18,5 +18,8 @@ def hashmodel(model, library = None):
     sha = hashlib.sha1(to_bytestring('{0}({1})'.format(library,meta)))
     hash = sha.hexdigest()[:8]
     meta.hash = hash
+    if hash in _model_dict:
+        raise KeyError('Model "{0}" already in hash table.\
+ Rename your model or the module containing the model.'.format(meta)) 
     _model_dict[hash] = model